src/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask==2.3.3
